Message timestamp refactoring and log level changes
