+++ include/cppkafka/message_timestamp.h
@@ -27,42 +27,47 @@
  *
  */
 
-#ifndef CPPKAFKA_H
-#define CPPKAFKA_H
+#ifndef CPPKAFKA_MESSAGE_TIMESTAMP_H
+#define CPPKAFKA_MESSAGE_TIMESTAMP_H
 
-#include <cppkafka/buffer.h>
-#include <cppkafka/clonable_ptr.h>
-#include <cppkafka/configuration.h>
-#include <cppkafka/configuration_base.h>
-#include <cppkafka/configuration_option.h>
-#include <cppkafka/consumer.h>
-#include <cppkafka/error.h>
-#include <cppkafka/exceptions.h>
-#include <cppkafka/group_information.h>
-#include <cppkafka/header.h>
-#include <cppkafka/header_list.h>
-#include <cppkafka/header_list_iterator.h>
-#include <cppkafka/kafka_handle_base.h>
-#include <cppkafka/logging.h>
-#include <cppkafka/macros.h>
-#include <cppkafka/message.h>
-#include <cppkafka/message_builder.h>
-#include <cppkafka/message_internal.h>
-#include <cppkafka/message_timestamp.h>
-#include <cppkafka/metadata.h>
-#include <cppkafka/producer.h>
-#include <cppkafka/queue.h>
-#include <cppkafka/topic.h>
-#include <cppkafka/topic_configuration.h>
-#include <cppkafka/topic_partition.h>
-#include <cppkafka/topic_partition_list.h>
-#include <cppkafka/utils/backoff_committer.h>
-#include <cppkafka/utils/backoff_performer.h>
-#include <cppkafka/utils/buffered_producer.h>
-#include <cppkafka/utils/compacted_topic_processor.h>
-#include <cppkafka/utils/consumer_dispatcher.h>
-#include <cppkafka/utils/poll_interface.h>
-#include <cppkafka/utils/poll_strategy_base.h>
-#include <cppkafka/utils/roundrobin_poll_strategy.h>
+#include <chrono>
+#include <boost/optional.hpp>
+#include <librdkafka/rdkafka.h>
+#include "macros.h"
 
-#endif
+namespace cppkafka {
+
+/**
+ * Represents a message's timestamp
+ */
+class CPPKAFKA_API MessageTimestamp {
+    friend class Message;
+public:
+    /**
+     * The timestamp type
+     */
+    enum TimestampType {
+        CREATE_TIME = RD_KAFKA_TIMESTAMP_CREATE_TIME,
+        LOG_APPEND_TIME = RD_KAFKA_TIMESTAMP_LOG_APPEND_TIME
+    };
+    
+    /**
+     * Gets the timestamp value. If the timestamp was created with a 'time_point',
+     * the duration represents the number of milliseconds since epoch.
+     */
+    std::chrono::milliseconds get_timestamp() const;
+
+    /**
+     * Gets the timestamp type
+     */
+    TimestampType get_type() const;
+private:
+    MessageTimestamp(std::chrono::milliseconds timestamp, TimestampType type);
+    
+    std::chrono::milliseconds timestamp_;
+    TimestampType type_;
+};
+
+} // cppkafka
+
+#endif //CPPKAFKA_MESSAGE_TIMESTAMP_H
